requirements.txt
Django==4.2.7
django-unfold==0.50.0
pymysql==1.1.2
python-dotenv==0.21.0
gunicorn==21.2.0
Pillow==10.1.0
